PHASE_07_SERVING_STREAMING/streaming/requirements.txt
kafka-python==2.0.2
requests==2.32.3
pymongo==4.8.0
